Backport Beta 2.2.1 (#9905)
* Bump beta to version 2.2.1

* fix: Intermittent failing CI due to addr in use (#9885)

Allow OS to set port at runtime

* Use Weak reference in PubSubClient (#9886)

* Fix json tracer overflow (#9873)

* Fix json tracer overflow

* Replace trace_executed with a direct trace push

* Remove unused variable

* Add test for 5a51

* Remove duplicate json!

* Fix docker script (#9854)


* Dockerfile: change source path of the newly added check_sync.sh (#9869)

* Allow to seal work on latest block (#9876)

* Allow to seal work on latest block.

* Test from @todr to check sealing conditions.

* gitlab-ci: make android release build succeed (#9743)

* use docker cargo config file for android builds

* make android build succeed

* ethcore: use Machine::verify_transaction on parent block (#9900)

* ethcore: use Machine::verify_transaction on parent block

also fixes off-by-one activation of transaction permission contract

* ethcore: clarify call to verify_transaction

* foundation: #6692865, ropsten: #4417537, kovan: #9363457

* Remove rust-toolchain file (#9906)

* EIP-712 implementation (#9631)

* EIP-712 impl

* added more tests

* removed size parsing unwrap

* corrected TYPE_REGEX to disallow zero sized fixed length arrays, replaced LinkedHashSet with IndexSet, added API spec to docs, fixed Type::Byte encoding branch

* use Option<u64> instead of u64 for Type::Array::Length

* replace `.iter()` with  `.values()`


* tabify eip712.rs

* use proper comments for docs

* Cargo.lock: revert unrelated changes

* tabify encode.rs

* EIP 191 (#9701)

* added sign_191 rpc method

* fixed hash_structured_data return type

* added ConfirmationPayload::SignMessage for non-prefixed signatures, added tests for sign191

* renamed WithValidator -> PresignedTransaction

* rename applicationData to data in test

* adds docs for EIP191Version, renamed SignRequest to EIP191SignRequest

* light-fetch: Differentiate between out-of-gas/manual throw and use required gas from response on failure (#9824)

* fix start_gas, handle OOG exceptions & NotEnoughGas

* Change START_GAS: 50_000 -> 60_000
* When the `OutOfGas exception` is received then try to double the gas until it succeeds or block gas limit is reached
* When `NotEnoughBasGas error` is received then use the required gas provided in the response

* fix(light-fetch): ensure block_gas_limit is tried

Try the `block_gas_limit` before regard the execution as an error

* Update rpc/src/v1/helpers/light_fetch.rs


* simplify cargo audit

* Use block header for building finality (#9914)

* ci: nuke the gitlab caches (#9855)

+++ scripts/gitlab/build-unix.sh
@@ -6,18 +6,23 @@ set -u # treat unset variables as error
 echo "__________Show ENVIROMENT__________"
 echo "CI_SERVER_NAME:   " $CI_SERVER_NAME
 echo "CARGO_HOME:       " $CARGO_HOME
-echo "BUILD_TARGET:     " $BUILD_TARGET
-echo "BUILD_ARCH:       " $BUILD_ARCH
 echo "CARGO_TARGET:     " $CARGO_TARGET
 echo "CC:               " $CC
 echo "CXX:              " $CXX
 
 echo "__________CARGO CONFIG__________"
-mkdir -p .cargo
-rm -f .cargo/config
-echo "[target.$CARGO_TARGET]" >> .cargo/config
-echo "linker= \"$CC\"" >> .cargo/config
-cat .cargo/config
+if [ "${CARGO_TARGET}" = "armv7-linux-androideabi" ]
+then
+  # use build container's cargo config
+  cat /.cargo/config
+else
+  mkdir -p .cargo
+  rm -f .cargo/config
+  echo "[target.$CARGO_TARGET]" >> .cargo/config
+  echo "linker= \"$CC\"" >> .cargo/config
+  cat .cargo/config
+fi
+
 
 echo "_____ Building target: "$CARGO_TARGET" _____"
 time cargo build --target $CARGO_TARGET --release --features final
@@ -32,15 +37,31 @@ mkdir -p artifacts
 cd artifacts
 mkdir -p $CARGO_TARGET
 cd $CARGO_TARGET
-cp ../../target/$CARGO_TARGET/release/parity ./parity
-cp ../../target/$CARGO_TARGET/release/parity-evm ./parity-evm
-cp ../../target/$CARGO_TARGET/release/ethstore ./ethstore
-cp ../../target/$CARGO_TARGET/release/ethkey ./ethkey
-cp ../../target/$CARGO_TARGET/release/whisper ./whisper
-strip -v ./*
+cp -v ../../target/$CARGO_TARGET/release/parity ./parity
+cp -v ../../target/$CARGO_TARGET/release/parity-evm ./parity-evm
+cp -v ../../target/$CARGO_TARGET/release/ethstore ./ethstore
+cp -v ../../target/$CARGO_TARGET/release/ethkey ./ethkey
+cp -v ../../target/$CARGO_TARGET/release/whisper ./whisper
+
+
+# stripping can also be done on release build time
+# export RUSTFLAGS="${RUSTFLAGS} -C link-arg=-s"
+if [ "${CARGO_TARGET}" = "armv7-linux-androideabi" ]
+then
+  arm-linux-androideabi-strip -v ./*
+else
+  strip -v ./*
+fi
+
 echo "_____ Calculating checksums _____"
 for binary in $(ls)
 do
   rhash --sha256 $binary -o $binary.sha256
-  ./parity tools hash $binary > $binary.sha3
+  if [ "${CARGO_TARGET}" = "armv7-linux-androideabi" ]
+  then
+    echo "> ${binary} cannot be hashed with cross-compiled binary"
+  else
+    ./parity tools hash $binary > $binary.sha3
+  fi
 done
+
